requirements.txt
asgiref==3.5.2
black==22.6.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
cloudinary==1.29.0
config==0.5.1
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==4.0.5
django-cors-headers==3.13.0
django-heroku==0.3.1
django-rest-swagger==2.2.0
djangorestframework==3.13.1
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
openapi-codec==1.3.2
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
psycopg2==2.9.3
pyparsing==3.0.9
python-decouple==3.6
pytz==2022.1
pyuca==1.2
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typing-extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
whitenoise==6.2.0
